chore(): fix validation import statements, Plugin=>EventsPlugin, added widget validation, moved plugin and widget interfaces to a separate file
