+++ ui/plugins/loader/src/validations.ts
@@ -1,21 +1,21 @@
-import { IPlugin, IWidget } from '.';
+import { IPlugin, IWidget } from './interfaces';
 
-const validateServices = services => {
+const validateServices = (services: any) => {
   return true;
 };
 
 export const validatePlugin = (plugin: IPlugin) => {
-  let isValid = false;
+  let isValid = true;
   if (Array.isArray(plugin.services)) {
     isValid = validateServices(plugin.services);
   }
-  return true;
+  return isValid;
 };
 
 export const validateWidget = (widget: IWidget) => {
-  let isValid = false;
+  let isValid = true;
   if (Array.isArray(widget.services)) {
     isValid = validateServices(widget.services);
   }
-  return true;
+  return isValid;
 };
